+++ main.go
@@ -53,7 +53,7 @@ var logger logging.LoggingInterface
 func getG2config(ctx context.Context) (g2api.G2config, error) {
 	result := g2config.G2config{}
 	moduleName := "Test module name"
-	verboseLogging := 0 // 0 for no Senzing logging; 1 for logging
+	verboseLogging := int64(0) // 0 for no Senzing logging; 1 for logging
 	iniParams := `{}`
 	err := result.Init(ctx, moduleName, iniParams, verboseLogging)
 	return &result, err
@@ -62,7 +62,7 @@ func getG2config(ctx context.Context) (g2api.G2config, error) {
 func getG2configmgr(ctx context.Context) (g2api.G2configmgr, error) {
 	result := g2configmgr.G2configmgr{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams := `{}`
 	err := result.Init(ctx, moduleName, iniParams, verboseLogging)
 	return &result, err
@@ -71,7 +71,7 @@ func getG2configmgr(ctx context.Context) (g2api.G2configmgr, error) {
 func getG2diagnostic(ctx context.Context) (g2api.G2diagnostic, error) {
 	result := g2diagnostic.G2diagnostic{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams := `{}`
 	err := result.Init(ctx, moduleName, iniParams, verboseLogging)
 	return &result, err
@@ -80,7 +80,7 @@ func getG2diagnostic(ctx context.Context) (g2api.G2diagnostic, error) {
 func getG2engine(ctx context.Context) (g2api.G2engine, error) {
 	result := g2engine.G2engine{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams := `{}`
 	err := result.Init(ctx, moduleName, iniParams, verboseLogging)
 	return &result, err
@@ -89,7 +89,7 @@ func getG2engine(ctx context.Context) (g2api.G2engine, error) {
 func getG2product(ctx context.Context) (g2api.G2product, error) {
 	result := g2product.G2product{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams := `{}`
 	err := result.Init(ctx, moduleName, iniParams, verboseLogging)
 	return &result, err
